feat(transit-card): 充值功能金额参数从分改为整数单位
修改充值金额的数据类型从decimal改为int，前端界面移除"(分)"单位显示，
后端接口参数统一调整为整数类型以简化金额处理逻辑。

BREAKING CHANGE: 充值API的amount参数现在接收整数而非小数

+++ src/DeviceHub.Service.Api/WebSocket/WebSocketHandler.cs
@@ -394,7 +394,7 @@ public sealed class WebSocketHandler : IDisposable
                 {
                     var readerName = GetParam(parameters, "readerName");
                     var amountStr = GetParam(parameters, "amount");
-                    if (!decimal.TryParse(amountStr, out var amount) || amount <= 0)
+                    if (!int.TryParse(amountStr, out var amount) || amount <= 0)
                     {
                         await SendErrorAsync(ws, requestId, "INVALID_PARAMETERS", "Invalid amount");
                         return;
